inital commit for stubbing out unit tests for ActivityController. Correcting message and handler file names that said "Command" to "Query" to reflect what they actually do.
