--- AllReadyApp/Web-App/AllReady/Features/Activity/GetMyActivitiesCommand.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using AllReady.ViewModels;
-using MediatR;
-
-namespace AllReady.Features.Activity
-{
-    public class GetMyActivitiesCommand : IRequest<MyActivitiesResultsScreenViewModel>
-    {
-        public string UserId { get; set; }
-    }
-}
