+++ AllReadyApp/Web-App/AllReady/Features/Activity/GetMyActivitiesQuery.cs
@@ -1,0 +1,10 @@
+﻿using AllReady.ViewModels;
+using MediatR;
+
+namespace AllReady.Features.Activity
+{
+    public class GetMyActivitiesQuery : IRequest<MyActivitiesResultsScreenViewModel>
+    {
+        public string UserId { get; set; }
+    }
+}
