--- AllReadyApp/Web-App/AllReady/Features/Activity/GetMyTasksCommand.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using AllReady.ViewModels;
-using MediatR;
-
-namespace AllReady.Features.Activity
-{
-    public class GetMyTasksCommand : IRequest<IEnumerable<TaskSignupViewModel>>
-    {
-        public int ActivityId { get; set; }
-        public string UserId { get; set; }
-    }
-}
