--- AllReadyApp/Web-App/AllReady/Features/Activity/ShowActivityCommand.cs
@@ -1,12 +1,0 @@
-using System.Security.Claims;
-using AllReady.ViewModels;
-using MediatR;
-
-namespace AllReady.Features.Activity
-{
-    public class ShowActivityCommand : IRequest<ActivityViewModel>
-    {
-        public int ActivityId { get; set; }
-        public ClaimsPrincipal User { get; set; }
-    }
-}
